[+] Add-Pipeline start refactor
