+++ src/components/Sidebar/SidebarRight/AddPipeline/Steps/index.js
@@ -1,0 +1,5 @@
+export { default as Initial } from './Initial.react';
+export { default as Nodes } from './Nodes.react';
+export { default as Webhooks } from './Webhooks.react';
+export { default as Options } from './Options.react';
+export { default as Triggers } from './Triggers.react';
