+++ src/config/schema/addPipeline.schema.js
@@ -1,0 +1,19 @@
+const addPipelineSchema = {
+  INITIAL: {
+    label: 'Initial'
+  },
+  NODES: {
+    label: 'Nodes'
+  },
+  WEBHOOKS: {
+    label: 'Webhooks'
+  },
+  TRIGGERS: {
+    label: 'Triggers'
+  },
+  OPTIONS: {
+    label: 'Options'
+  }
+};
+
+export default addPipelineSchema;
