+++ src/styles/Display.styles.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+const Display = styled.div`
+  display: ${({ isVisible }) => (isVisible ? 'block' : 'none')};
+`;
+
+export default Display;
